feat/Incrementado o service e feito o controller

+++ src/main/java/com/alquimia/backend/service/ItemPedidoService.java
@@ -4,6 +4,7 @@ import com.alquimia.backend.dto.request.ItemPedidoRequestDTO;
 import com.alquimia.backend.dto.response.ItemPedidoResponseDTO;
 import com.alquimia.backend.model.ItemEstoque;
 import com.alquimia.backend.model.ItemPedido;
+import com.alquimia.backend.model.Pedido;
 import com.alquimia.backend.model.Produto;
 import com.alquimia.backend.repository.ItemEstoqueRepository;
 import com.alquimia.backend.repository.ItemPedidoRepository;
@@ -15,6 +16,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ItemPedidoService {
@@ -47,23 +49,55 @@ public class ItemPedidoService {
     }
 
     @Transactional
-    public void  removerItemDoPedido(Integer cdItemPedido) {
+    public Integer removerItemDoPedido(Integer cdItemPedido) {
         ItemPedido itemPedido = itemPedidoRepository.findAllByCdItemPedido(cdItemPedido)
-                .orElseThrow(()-> new RuntimeException("Item do pedido não encontrado."));
+                .orElseThrow(() -> new RuntimeException("Item do pedido não encontrado."));
 
         int qtItemPedido = itemPedido.getQtItemPedido();
-        itemEstoque.aumentarQtdeItemEstoque(qtItemPedido);
+        Produto produto = itemPedido.getCdProduto();
+        Integer cdPedidoAtual = itemPedido.getCdPedido().getCdPedido();
 
+        Optional<ItemEstoque> itemEstoqueOptional = itemEstoqueRepository.findByCdProduto(produto.getCdProduto());
+        if (itemEstoqueOptional.isEmpty()) {
+            throw new RuntimeException("Estoque não encontrado para o produto: " + produto.getCdProduto());
+        }
+
+        ItemEstoque itemEstoque = itemEstoqueOptional.get();
+        itemEstoque.setQtItemEstoque(itemEstoque.getQtItemEstoque() + qtItemPedido);
+        itemEstoqueRepository.save(itemEstoque);
         itemPedidoRepository.delete(itemPedido);
+
+        return cdPedidoAtual;
     }
 
-    public List<ItemPedidoResponseDTO> listarItensPedido(Integer cdItemPedido) {
+    public List<ItemPedidoResponseDTO> listarTodosItensPedido() {
         List<ItemPedidoResponseDTO> listaItensPedidoDTO = new ArrayList<>();
-        List<ItemPedido> pedido = itemPedidoRepository.findAll();
-        for (ItemPedido itemPedido : pedido) {
+        List<ItemPedido> itens = itemPedidoRepository.findAll();
+        for (ItemPedido itemPedido : itens) {
             listaItensPedidoDTO.add(new ItemPedidoResponseDTO(itemPedido));
         }
         return listaItensPedidoDTO;
     }
 
+    @Transactional
+    public void reverterEstoqueDoPedido(Pedido pedido) {
+        List<ItemPedido> itensParaReverter = itemPedidoRepository.findAllByCdPedido(pedido);
+
+        if (itensParaReverter.isEmpty()) {
+            return;
+        }
+
+        for (ItemPedido itemPedido : itensParaReverter) {
+            int qtItemPedido = itemPedido.getQtItemPedido();
+            Produto produto = itemPedido.getCdProduto();
+
+            Optional<ItemEstoque> itemEstoqueOptional = itemEstoqueRepository.findByCdProduto(produto.getCdProduto());
+
+            ItemEstoque itemEstoque = itemEstoqueOptional.get();
+
+            itemEstoque.setQtItemEstoque(itemEstoque.getQtItemEstoque() + qtItemPedido);
+            itemEstoqueRepository.save(itemEstoque);
+        }
+
+    }
 }
